feat: exception and errors in user (user done)

+++ src/users/users.controller.ts
@@ -6,37 +6,56 @@ import {
   Patch,
   Param,
   Delete,
+  HttpCode,
+  BadRequestException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-
+import { validate } from 'uuid';
 @Controller('user')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
+  @HttpCode(201)
   create(@Body() createUserDto: CreateUserDto) {
     return this.usersService.create(createUserDto);
   }
 
   @Get()
+  @HttpCode(200)
   findAll() {
     return this.usersService.findAll();
   }
 
   @Get(':id')
+  @HttpCode(200)
   findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+    if (validate(id)) {
+      return this.usersService.findOne(id);
+    } else {
+      throw new BadRequestException();
+    }
   }
 
   @Patch(':id')
+  @HttpCode(200)
   update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
-    return this.usersService.update(id, updateUserDto);
+    if (validate(id)) {
+      return this.usersService.update(id, updateUserDto);
+    } else {
+      throw new BadRequestException();
+    }
   }
 
   @Delete(':id')
+  @HttpCode(204)
   remove(@Param('id') id: string) {
-    return this.usersService.remove(id);
+    if (validate(id)) {
+      return this.usersService.remove(id);
+    } else {
+      throw new BadRequestException();
+    }
   }
 }
